V 1.1

+++ src/app/register/register.page.ts
@@ -9,6 +9,7 @@ import { Router } from '@angular/router';
 export class RegisterPage {
   newUsername = ''; // Nombre de usuario a registrar
   newPassword = ''; // Contraseña a registrar
+  message = ''; // Mensaje para mostrar al usuario
 
   registeredUsers: { username: string, password: string }[] = [];
 
@@ -20,16 +21,24 @@ export class RegisterPage {
   }
   
   register() {
-    // Agregar el nuevo usuario registrado al arreglo
-    this.registeredUsers.push({ username: this.newUsername, password: this.newPassword });
-  
-    // Almacenar el arreglo en localStorage
-    localStorage.setItem('registeredUsers', JSON.stringify(this.registeredUsers));
-  
-    // Redirigir a la página de inicio de sesión
-    this.router.navigate(['/login']);
+    // Verificar si el nombre de usuario ya existe
+    const userExists = this.registeredUsers.some(user => user.username === this.newUsername);
+
+    if (userExists) {
+      this.message = 'El nombre de usuario ya existe';
+    } else {
+      // Agregar el nuevo usuario registrado al arreglo
+      this.registeredUsers.push({ username: this.newUsername, password: this.newPassword });
+    
+      // Almacenar el arreglo en localStorage
+      localStorage.setItem('registeredUsers', JSON.stringify(this.registeredUsers));
+    
+      // Redirigir a la página de inicio de sesión
+      this.router.navigate(['/login']);
+    }
   }
 }
+
